+++ extension/src/extension.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { findTranslationsInALFile, findTranslationsInXliffFile } from './ATCFunctions';
+import { writeTranslationsToCommentsInALFile, writeTranslationsFromXliffToALFile } from './ATCFunctions';
 import { setLogger } from './logging/LogHelper';
 import { OutputLogger } from './logging/OutputLogger';
 
@@ -8,8 +8,8 @@ export function activate(context: vscode.ExtensionContext) {
     console.log('Extension al-translation-cleaner activated');
 
     context.subscriptions.push(
-        vscode.commands.registerCommand('atc.findTranslationsInALFile', findTranslationsInALFile),
-        vscode.commands.registerCommand('atc.findTranslationsInXliffFile', findTranslationsInXliffFile)
+        vscode.commands.registerCommand('atc.writeTranslationsToCommentsInALFile', writeTranslationsToCommentsInALFile),
+        vscode.commands.registerCommand('atc.writeTranslationsFromXliffToALFile', writeTranslationsFromXliffToALFile)
     );
 
 }